New script to update filter documents
