+++ README.md
@@ -1,3 +1,7 @@
 dp-data-tools-go
 ==================
 Contains several various tools to update data
+
+Current tools/scripts
+* [Edition document restructure](./mongo-fixes/edition-doc-structure)
+* [Filter blueprint and output documents include new dataset object](./mongo-fixes/filter-doc-version-identifier)
