+++ mongo-fixes/edition-doc-structure/README.md
@@ -1,0 +1,12 @@
+edition-doc-structure
+==================
+
+This script is to update edition resources on cmd due to structural changes.
+
+### How to run service
+* Run `go build`
+* Run `./edition-doc-structure -mongo-url=<url>`
+
+The url should look something like the following `localhost:27017` or
+`127.0.0.1:27017`. If a username and password are needed follow this structure
+`<username>:<password>@<host>:<port>`
